refactor(cli-better-auth): address PR #243 review — types.ts, shared URL consts, doc cleanups (PR #243)
- Move auth interfaces (AuthInputs/ResolvedAuth, DeviceCode/Token, SessionUser,
  PollOptions) into src/auth/types.ts; resolve.ts/device.ts/client.ts import them.
- Hoist DEFAULT_DASHBOARD_URL/DEFAULT_BASE_URL into src/constants.ts; login +
  project-id import from there (no more per-file copies).
- Rename credentialsDir -> globalLmnrDirectory (config dir may hold more later);
  trim its readCredentials comment.
- Drop stale LMNR_PROJECT_ID from the sql --project-id help.
- README: remove the removed `traces wait` command + LMNR_PROJECT_ID; fix
  --project -> --project-id.
