+++ packages/lmnr-cli/src/constants.ts
@@ -1,0 +1,5 @@
+// Default Laminar endpoints, shared across commands. Override per-invocation
+// via the --dashboard-url / --base-url flags or the LMNR_DASHBOARD_URL /
+// LMNR_BASE_URL env vars; these are the final fallback.
+export const DEFAULT_DASHBOARD_URL = "https://www.laminar.sh";
+export const DEFAULT_BASE_URL = "https://api.lmnr.ai";
